refactor: OrderBook 제거 및 OrderMenu 수정 - OrderBook 클래스에서 메뉴 아이템의 유효성 검사를 수행했는데, MenuItem을 구현한 enum들과 Menu에서 수행하도록 변경. - OrderBook 클래스가 OrderMenu와 역할이 비슷하기 때문에 OrderBook에 있는 메서드들을 OrderMenu로 이동하고 OrderBook 클래스 제거. - MenuItem 인터페이스에 검증시 필요한 메서드 추가.
