+++ src/main/java/christmas/domain/enums/Appetizer.java
@@ -1,5 +1,7 @@
 package christmas.domain.enums;
 
+import static christmas.exception.ExceptionMessage.INVALID_ORDER_MENU;
+
 public enum Appetizer implements MenuItem {
     MUSHROOM_SOUP("양송이수프", 6_000),
     TAPAS("타파스", 5_500),
@@ -22,5 +24,18 @@ public enum Appetizer implements MenuItem {
     @Override
     public int getPrice() {
         return this.price;
+    }
+
+    @Override
+    public boolean isSameName(String name) {
+        validate(name);
+        return this.name.equals(name);
+    }
+
+    @Override
+    public void validate(String name) {
+        if (name == null || name.isEmpty()) {
+            throw new IllegalArgumentException(INVALID_ORDER_MENU.getMessage());
+        }
     }
 }
