+++ src/main/java/christmas/domain/enums/Dessert.java
@@ -1,5 +1,7 @@
 package christmas.domain.enums;
 
+import static christmas.exception.ExceptionMessage.INVALID_ORDER_MENU;
+
 public enum Dessert implements MenuItem {
     CHOCOLATE_CAKE("초코케이크", 15_000),
     ICE_CREAM("아이스크림", 5_000),
@@ -21,5 +23,18 @@ public enum Dessert implements MenuItem {
     @Override
     public int getPrice() {
         return this.price;
+    }
+
+    @Override
+    public boolean isSameName(String name) {
+        validate(name);
+        return this.name.equals(name);
+    }
+
+    @Override
+    public void validate(String name) {
+        if (name == null || name.isEmpty()) {
+            throw new IllegalArgumentException(INVALID_ORDER_MENU.getMessage());
+        }
     }
 }
