+++ src/main/java/christmas/domain/enums/MainCourse.java
@@ -1,5 +1,7 @@
 package christmas.domain.enums;
 
+import static christmas.exception.ExceptionMessage.INVALID_ORDER_MENU;
+
 public enum MainCourse implements MenuItem {
     T_BONE_STEAK("티본스테이크", 55_000),
     BBQ_RIB("바비큐립", 54_000),
@@ -23,5 +25,18 @@ public enum MainCourse implements MenuItem {
     @Override
     public int getPrice() {
         return this.price;
+    }
+
+    @Override
+    public boolean isSameName(String name) {
+        validate(name);
+        return this.name.equals(name);
+    }
+
+    @Override
+    public void validate(String name) {
+        if (name == null || name.isEmpty()) {
+            throw new IllegalArgumentException(INVALID_ORDER_MENU.getMessage());
+        }
     }
 }
