+++ src/main/java/christmas/domain/enums/MenuItem.java
@@ -4,4 +4,8 @@ public interface MenuItem {
     String getName();
 
     int getPrice();
+
+    boolean isSameName(String name);
+
+    void validate(String name);
 }
